refactor(server): extract recommend_skills to its own SkillService impl block
FU-3 partial of v0.8.0 refinement. The 159-LOC `recommend_skills`
method was the largest single method on `SkillService` — dependency
walking + sibling detection + quality scoring with three nearly
identical loops. Move to `app/skill_recommendations.rs` as a
dedicated `impl SkillService` block.

Rust allows multiple `impl SkillService { ... }` blocks across
files; the split is mechanical and the public API is preserved.

app/mod.rs:    731 -> 567 LOC (-22% in this commit, -58% from
              the original 1,359 LOC pre-T3.1)
app/skill_recommendations.rs: 182 LOC (new, with module docstring)

The `compute_metrics` (155 LOC) and `read_resource_sync` (72 LOC)
methods remain in mod.rs as backlog candidates for the same
treatment. They are smaller and lower-priority since they're
already organized as topical regions; further extraction is
incremental.

480 server-lib tests pass.
